Use wordpress language to determine disqus lang
People use multisite install with different languages. Each site should load disqus in the language it is defined in.

Additional checking could be done on the $wplang variable to see if it is available at disqus and if not default to the disqus configuration one.

A better solution would be to offer a dropdown in the disqus wordpress settings page to chose the language. All problem would them be solved.